Added basic css to all webpages
Added new font-families, colors, and background-
colors to all the different webpages. Used types
for features I wanted for all webpages and
classes for stuff like ingredients and
instructions which are on the recipes but not the
homepage. "Fried Rice" had the most touchups, but
the rest of the pages will get work as well.